Audio to text/requirements.txt
openai-whisper
ffmpeg
torch cuda
